Added ES version to index names
ES does not behave well when documents from different versions are
stored in the same index.

So, to prevent errors, it's good to have different indices to store
documents from different ES versions.

Also, to downgrade we may need to delete the documents from the version
above, so with different indices this is easier.