Deprecated some logstash configuration folders
The folders `cf-app-es`, `cf-app-es-mirroring` and
`cf-apps-es-throttling-stdout-debug` are already a few changes behing
the configuration used on the application GCP cluster:
`cf-apps-es-throttling`

In order to clean up the repo a bit, we shouldn't have deprecated
configurations hanging around here. If someone wants to use them (for
any reason), they will have breaking changes, so they are not ready for
PROD.

We should _mark_ them as deprecated and delete them in the future.