Added a filter to set proper index level for some cases
Some apps send their log level as integers: 20000 for info, 30000 for
warn and 40000 for error.

We are using the level field in the index name. This means that this
apps are creating indices on ES just because of the level field being
different than all the other apps (they send info|error|warn)

With this new logstash filter, we change the index level value from the
integer to the proper value, thus creating less indices and improving
the performance of the cluster