Added sufix to index names
According to the oficials docs:
```
The rollover API supports date math, but requires the index name to end with a dash followed by a number, e.g. logstash-2016.02.03-1 which is incremented every time the index is rolled over.
```
https://www.elastic.co/guide/en/elasticsearch/reference/6.5/indices-rollover-index.html

We want to try the rollover API, so we need to change the index name
first. A sufix was added to be complient with what the oficials docs
say.