Changed logstash output to write to alias
Before we had logstash writing to ES indices based on the index name.

We now use aliases on ES, so we need to tell logstash to write to the
alias instead.

For example, logstash will write to the alias `logs-info` which will be
pointing to an active index on ES.

https://app.clubhouse.io/springernature/story/4522/implement-rollover-indices