Added input to port 5515 - rtr opt in
In the context of
https://app.clubhouse.io/springernature/story/5019/implement-rtr-opt-in,
we are adding a new input to logstash that listens on port 5515.

We then add a field called `[rtr][enabled]` via that input that will
allow us, in the future, to drop all RTR logs that are _not_ sent to
port 5515.

For now the dropping feature is commented out because it's a breaking
change and we need to give time to developers to make the switch