Added drain script that will wait for empty buffer
Logstash can buffer events into disc. In this case, we may not want
logstash to be deployed because we can lose the data in the buffer (for
example, if we remove an instance that still has files on buffer).

With the drain script, the deployment will wait until the buffer is
empty before succeeding. In case it's not successful  on time, it will
fail.

We're checking for 2 (or more) `page*` files because we saw that there
is always a `page` file on disc, even if the buffer is empty. However,
if there are 2 files, it means that the buffer is NOT empty.